requirements.txt
# Core dependencies
# Note: Install openai-whisper separately using: pip install -U openai-whisper
# openai-whisper>=20231106
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.1.0
ffmpeg-python>=0.2.0

# GUI
tkinter-tooltip>=2.1.0

# Utilities
python-dotenv>=1.0.0
